requirements.txt
prometheus-client
python-swiftclient
python-json-logger==0.1.5
logging
